Partly revert an earlier commit: don't use class methods for mask/shading images and filters, as there may be several presentations on different screens (really).
Set default value for (cached) filter. In particular the start time should be reset to 0. 

Shift the location of the mask image and of the swipe extent. 

Disable flushWindow of the fullscreen window during animation, so we can be sure the next state is not shown on screen too early.

git-svn-id: http://svn.code.sf.net/p/skim-app/code/trunk@2517 0fceea05-a30d-0410-8a8b-80ef821fb0a1